# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Capacitarr API
  version: 1.0.0
  description: |
    Capacitarr is a capacity management tool for media servers. It monitors disk
    usage across Sonarr, Radarr, Lidarr, Readarr, Plex, Jellyfin, Emby, Tautulli,
    and Overseerr integrations, scores media items for potential cleanup based on
    configurable weights and rules, and can automatically (or via approval) free
    disk space when thresholds are breached.

    ## Authentication

    Most endpoints require authentication via one of:
    - **JWT cookie** (`jwt`) — set automatically on login
    - **Bearer token** — `Authorization: Bearer <token>` header
    - **API key** — `X-Api-Key: <key>` header

    The `/health`, `/version`, and `/auth/login` endpoints are public.
  license:
    name: PolyForm Noncommercial 1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0/
servers:
  - url: http://localhost:2187/api/v1
    description: Local development server
security:
  - cookieAuth: []
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /rule-fields:
    get:
      operationId: getRuleFields
      tags: [Rules]
      summary: Get available rule fields
      description: |
        Returns the list of available fields for building protection/targeting
        rules. Fields are dynamic based on which integrations are enabled.
        For example, Sonarr-specific fields (season count, episode count) only
        appear when Sonarr is enabled.
      parameters:
        - name: service_type
          in: query
          description: Filter fields for a specific integration type
          schema:
            type: string
            enum: [sonarr, radarr, lidarr, readarr]
      responses:
        "200":
          description: List of rule fields
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RuleField"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /rule-values:
    get:
      operationId: getRuleValues
      tags: [Rules]
      summary: Get autocomplete values for a rule field
      description: |
        Returns value options for a given rule field and integration. Depending
        on the field, returns either a closed set of options, a combobox with
        suggestions, or free-form input metadata. Results are cached for 5 minutes.
      parameters:
        - name: integration_id
          in: query
          required: true
          description: ID of the integration to fetch values from
          schema:
            type: integer
        - name: action
          in: query
          required: true
          description: |
            The rule field name to get values for (e.g., `quality`, `tag`,
            `genre`, `language`, `title`, `rating`, `sizebytes`, `timeinlibrary`,
            `year`, `monitored`, `requested`, `type`, `seriesstatus`,
            `seasoncount`, `episodecount`, `playcount`, `requestcount`)
          schema:
            type: string
      responses:
        "200":
          description: |
            Rule value options. Shape varies by field type:
            - **closed**: `{ type: "closed", options: NameValue[] }`
            - **combobox**: `{ type: "combobox", suggestions: NameValue[] }`
            - **free**: `{ type: "free", inputType: string, placeholder: string, suffix: string }`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleValueResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Integration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # Custom Rules
  # ===========================================================================
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key generated via `POST /auth/apikey`
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login response
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: HttpOnly JWT cookie set on login

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: Bad request — invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required: [error]
      example:
        error: "Invalid request body"
...
    RuleField:
      type: object
      properties:
        field:
          type: string
          example: quality
        label:
          type: string
          example: "Quality Profile"
        type:
          type: string
          enum: [string, number, boolean]
        operators:
          type: array
          items:
            type: string
          example: ["==", "!="]
      required: [field, label, type, operators]
    RuleValueResult:
      type: object
      description: |
        Polymorphic response — shape depends on the field type:
        - **closed**: `{ type: "closed", options: NameValue[] }`
        - **combobox**: `{ type: "combobox", suggestions: NameValue[] }`
        - **free**: `{ type: "free", inputType: string, placeholder: string, suffix: string }`
      properties:
        type:
          type: string
          enum: [closed, combobox, free]
        options:
          type: array
          items:
            $ref: "#/components/schemas/NameValue"
          description: Present when type is "closed"
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/NameValue"
          description: Present when type is "combobox"
        inputType:
          type: string
          description: Present when type is "free" (e.g., "text", "number")
        placeholder:
          type: string
          description: Present when type is "free"
        suffix:
          type: string
          description: Present when type is "free" (e.g., "days", "bytes")
      required: [type]
    NameValue:
      type: object
      properties:
        value:
          type: string
        label:
          type: string
      required: [value, label]
      example:
        value: "4K"
        label: "4K"

    # ---- Media / Scoring ----
...